# backend/src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EM TaskFlow AI - API Reference",
    "version": "1.0.0",
    "description": "Interactive API specification for the EM TaskFlow AI polyglot microservices platform. Powers local multi-agent supervisor reasoning (Ollama hermes3:8b), Hybrid RAG (HyDE + RRF), multi-source MCP integrations (Jira OAuth PKCE, Notion, GitHub, Slack, Google Calendar), and multi-session persistence."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/chat": {
      "post": {
        "tags": ["Chat & Multi-Agent"],
        "summary": "Execute Chat Inference",
        "description": "Processes user query via Redis semantic cache, Fast-Path (<300ms), single-pass RAG synthesizer, or LangGraph multi-agent supervisor.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message"],
                "properties": {
                  "message": { "type": "string", "example": "Summarize Q3 sprint blockers and DORA deployment frequency" },
                  "threadId": { "type": "string", "example": "th_9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d" },
                  "mode": { "type": "string", "enum": ["baseline", "advanced"], "default": "baseline" },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "filename": { "type": "string", "example": "sprint_review.pdf" },
                        "content": { "type": "string" },
                        "mimeType": { "type": "string", "example": "application/pdf" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Structured Markdown response generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": { "type": "string", "example": "### 📄 Executive Summary\n..." },
                    "threadId": { "type": "string" },
                    "domains": { "type": "array", "items": { "type": "string" } },
                    "cached": { "type": "boolean" },
                    "latencyMs": { "type": "number" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid chat payload" }
        }
      }
    },
    "/api/sessions": {
      "get": {
        "tags": ["Session & Thread"],
        "summary": "List Paginated Sessions",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "List of active client sessions with thread titles and timestamps"
          }
        }
      },
      "post": {
        "tags": ["Session & Thread"],
        "summary": "Create New Session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "default": "New Chat" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Session created successfully" }
        }
      }
    },
...
    "/api/threads": {
      "post": {
        "tags": ["Session & Thread"],
        "summary": "Create Thread in Current Session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "default": "New Chat" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Thread created" }
        }
      }
    },
    "/api/sessions/{sessionId}/switch": {
      "post": {
        "tags": ["Session & Thread"],
        "summary": "Switch Active Session Thread",
        "parameters": [
          { "name": "sessionId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["threadId"],
                "properties": {
                  "threadId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Active thread switched" }
        }
      }
    },
    "/api/threads/{threadId}/messages": {
      "get": {
        "tags": ["Session & Thread"],
        "summary": "Get Thread Message History",
        "parameters": [
          { "name": "threadId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": { "description": "Chronological messages list" }
        }
      }
    },
...
    "/api/feedback": {
      "post": {
        "tags": ["Chat & Multi-Agent"],
        "summary": "Submit User Feedback",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["score"],
                "properties": {
                  "score": { "type": "string", "enum": ["thumbs_up", "thumbs_down"] },
                  "threadId": { "type": "string" },
                  "traceId": { "type": "string" },
                  "comment": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Feedback submitted asynchronously" }
        }
      }
    }
  }
}